requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary  # For PostgreSQL
gunicorn  # For deployment